src/austerus-core.c,h: Sanitise naming and handling of buffers.
This patch renames the buffers pipe_buffer and serial_buffer to
line_gcode line_feedback respectively to be in line with the pipe and
stream naming used in austerus-panel.c.

The creation of the buffers has been improved and they are now freed
correctly.

This patch also makes blank lines from stdin ignored rather than sending
them on the serial port.