src/serial.c,h,austerus-core.c,h: Make serial read block correctly
This patch fixes a fundamental flaw in the serial communications code.

Previously the serial port was opened with the "O_NDELAY" parameter
which was causes read() not to block even when VMIN & VTIME were set
appropriately.

This patch completely refactors serial_getline() and introduces the
timeout argument which is given in seconds.

Now that we have solid blocking serial communications with configurable
timeout to build on; all of the "serial retries" and sleep nonsense has
been removed from austerus-core.

In austerus-core we are currently using a timeout of 10s for all calls
to serial_getline(). Which seems like a good balance between waiting
long enough and potential damage to prints.