src/austerus-core.c,h: Add basic internal command system
In a parent program calling austerus-core we will sometimes want to give
austerus-core (not the printer) instructions that need to be carried out
at a particular point in the gcode stream.

The example is when the parent program wants to exit only when
austerus-core has completely finished sending all gcode. To do this we
can put the new "#ag:exit" command into the end of the gcode stream and
then block until austerus-core exits upon processing this command.

All commands are prexied with "#ag:". Currently the only supported
command is  "#ag:exit".