src/austerus-send.c,h: Add austerus-send program
This patch adds the austerus-send program which is used for sending
gcode from files. Each file to be transmitted is given as an argument
to austerus-send and they are sent in sequence.

The program used austerus-core so the sending of gcode is fully
buffered and will never wait on disk.

This is the bare minimum implementation required to print and most
features which justify it's existence over just "cat [filename] |
austerus-core" are still to be implemented.