src/austerus-core,send: Move core from options to env variables
The core program is mostly intended to be called from other programs
and not generally exposed to the user. By using environmental
variables instead of command options the setting of options is more
flexible.

austerus-send now checks that serial port is either given as an option
or the environmental variable is set as this is required. All other
variables have default values found in defaults.h.