src/austerus-send.c: Implement basic progress stats
This patch implements a basic print progress display on the console.

The file in read and an array is built up holding the extruded length at
each line in the file. Using the total extruded length and this array we
can calculate the progress at any line in the print.

This patch is incomplete as it is using the number of lines sent down the
pipe rather than the number of ACK'd lines back from austerus-core to
calculate progress so it will always state a progress ahead by the number
of lines held in the pipe. To fix this we need to use popen2().