Capture all austerus-core output
Previously we were closing the feedback pipe before austerus-core had
finished writing all feedback to it. This meant that progress never
reached 100% when the printing of a file was completed.

This patch is not perfect as it is not removing the unblocking from the
feedback pipe, but this is a tricky problem to solve.