Add physical mode (-d) to fix test case "default-shifted"
This patch adds a new mode that takes offset caused by G92 "set" into
account. When set is used the values of the axes being tracked may be
different to the physical location of the machine.

When using the physical option the extent values output will represent the
actual physical positions of the machine, not the values found in the
gcode file.