Add austerus-verge --zmin mode
This mode returns the bounds of the region in which the Z axis is less
than zmin. Once we have determined a safe working Z height from existing
depositions on plate we can find the region in which a new piece of gcode
violates this (so we can check this region is clear).